+++ CHANGES.rst
@@ -12,6 +12,7 @@ lqdn-gafam-campaign
 - Publish/overwrite Arabic resources as they look more beautiful when rendered on Mac OS X
 - Improve introductional text on chooser page
 - Create 10x8 mosaic from all languages
+- Add SVG rendering
 
 
 2018-05-13 0.7.0